requirements.txt
# Core API runtime
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
gunicorn>=22.0,<24.0
slowapi>=0.1.9,<1.0
pydantic>=2.7,<3.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0

# Math + AI
sympy>=1.13,<2.0
groq>=0.11,<1.0

# RAG / embeddings
qdrant-client>=1.10,<2.0
sentence-transformers>=3.0,<4.0
langchain-text-splitters>=0.2,<1.0
pypdf>=4.3,<6.0
pdfplumber>=0.11,<1.0

# Supabase integration (scripts + APIs)
supabase>=2.6,<3.0

# Data ingestion/scraping/cleaning utilities
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
selenium>=4.24,<5.0
webdriver-manager>=4.0,<5.0
Pillow>=10.4,<12.0
wordninja>=2.0,<3.0
